[유호근] 관리자 페이지 해당 페이지, 상세보기로 가는 경로 설정 완료
