+++ src/main/java/com/app/neos/repository/admin/AdminInquiryRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface AdminInquiryRepository extends JpaRepository<Inquiry, Long> {
+//    한 유저가 쓴 모든 문의 가져오기
     public List<Inquiry> findByUser_UserId(Long userId);
 }
