+++ src/main/java/com/app/neos/repository/admin/AdminStoreReplyRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface AdminStoreReplyRepository extends JpaRepository<StoreReply, Long> {
+//    한 유저가 쓴 모든 자료 상점 댓글 가져오기
     public List<StoreReply> findByUser_UserId(Long userId);
 }
