+++ src/main/java/com/app/neos/repository/admin/AdminStudyFeedReplyRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 public interface AdminStudyFeedReplyRepository extends JpaRepository<StudyFeedReply, Long> {
-
+// 한 유저가 쓴 모든 피드 댓글 가져오기
     public List<StudyFeedReply> findAllByStudyFeedReplyWriter_UserId(Long userId);
 
 }
