+++ src/main/java/com/app/neos/service/admin/AdminService.java
@@ -173,75 +173,6 @@ public class AdminService {
     }
 
 
-//    //    유저 목록 페이지
-//    public Page<UserDTO> findAllUserPage(Pageable pageable){
-//        List<UserDTO> userDTOS = jpaQueryFactory.select(new QUserDTO(
-//                user.userId,
-//                user.userNickName,
-//                user.userOAuthId,
-//                user.userOAuthEmail,
-//                user.userCollegeEmail,
-//                user.userPhoneNumber,
-//                user.userCollegeCertify,
-//                user.userCollegeInfo.userCollegeYear,
-//                user.userCollegeInfo.userCollegeMajor,
-//                user.userNeosPower.userNeosBadge,
-//                user.userNeosPower.userNeosPowerLevel,
-//                user.userNeosPower.userNeosPowerAbility,
-//                user.userNeosPoint,
-//                user.userChattingPoint,
-//                user.userLike.userO2o,
-//                user.userLike.userCity,
-//                user.userLike.userDay,
-//                user.userLike.userTime,
-//                user.userMBTI.userMbtiName,
-//                user.userMBTI.userMbtiColor,
-//                user.userIntroduce,
-//                user.userFile,
-//                user.createdDate
-//        ))
-//                .from(user)
-//                .orderBy(user.userId.desc())
-//                .offset(pageable.getOffset())
-//                .limit(pageable.getPageSize())
-//                .fetch();
-//
-//        for (UserDTO user: userDTOS) {
-//            user.setCounts(adminService.workCount(user.getUserId()));
-//        }
-//
-//
-//        long total = jpaQueryFactory.select(new QUserDTO(
-//                user.userId,
-//                user.userNickName,
-//                user.userOAuthId,
-//                user.userOAuthEmail,
-//                user.userCollegeEmail,
-//                user.userPhoneNumber,
-//                user.userCollegeCertify,
-//                user.userCollegeInfo.userCollegeYear,
-//                user.userCollegeInfo.userCollegeMajor,
-//                user.userNeosPower.userNeosBadge,
-//                user.userNeosPower.userNeosPowerLevel,
-//                user.userNeosPower.userNeosPowerAbility,
-//                user.userNeosPoint,
-//                user.userChattingPoint,
-//                user.userLike.userO2o,
-//                user.userLike.userCity,
-//                user.userLike.userDay,
-//                user.userLike.userTime,
-//                user.userMBTI.userMbtiName,
-//                user.userMBTI.userMbtiColor,
-//                user.userIntroduce,
-//                user.userFile,
-//                user.createdDate
-//        ))
-//                .from(user)
-//                .fetch().size();
-//
-//        return new PageImpl<>(userDTOS,pageable,total);
-//    }
-
     //    유저 아이디로 조회
     public User findByUserId(Long userId){
         return  adminCustomRepository.findByUserId(userId);
