+++ src/main/resources/templates/app/admin/adminNavBar.html
@@ -44,7 +44,7 @@
                 <li class="spacer"></li> <!-- 공간 여백 div -->
 
                 <li class="active expanded expanding" data-title="adminPage">
-                    <a th:href="@{/admin/test/index}"> <!-- 관리자 메인 화면 링크-->
+                    <a th:href="@{/admin/index}"> <!-- 관리자 메인 화면 링크-->
                         <div class="gui-icon">
                             <i class="db-dashboard"></i>
                         </div>
@@ -61,12 +61,12 @@
                     </a>
                     <ul>
                         <li class="  " data-title="memberList">
-                            <a th:href="@{/admin/test/user/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/user/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title">사용자 목록</span>
                             </a>
                         </li>
                         <li class="  " data-title="companyList">  <!-- a태그의 클래스명이 li 태그 클래스에 출가된다. 활성화 된 클래스 표시(active) -->
-                            <a th:href="@{/admin/test/college/list}">  <!-- 해당 페이지의 목록에 active가 들어가야지 음영처리가된다. -->
+                            <a th:href="@{/admin/college/list}">  <!-- 해당 페이지의 목록에 active가 들어가야지 음영처리가된다. -->
                                 <span class="title">대학교 목록</span>
                             </a>
                         </li>
@@ -81,22 +81,22 @@
                     </a>
                     <ul>
                         <li class="  " data-title="studyList">
-                            <a th:href="@{/admin/test/study/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/study/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title">스터디 게시글 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="freeList">
-                            <a th:href="@{/admin/test/community/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/community/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title">자유 게시글 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="counselingList">
-                            <a th:href="@{/admin/test/counseling/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/counseling/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title">고민 상담 게시글 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="counselingList">
-                            <a th:href="@{/admin/test/store/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/store/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title">자료 상점 게시글 관리</span>
                             </a>
                         </li>
@@ -112,22 +112,22 @@
                     </a>
                     <ul>
                         <li class="  " data-title="studyList">
-                            <a  th:href="@{/admin/test/study/reply/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a  th:href="@{/admin/study/reply/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title">스터디 댓글 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="freeList">
-                            <a th:href="@{/admin/test/community/reply/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/community/reply/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title" >자유 게시판 댓글 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="counselingList">
-                            <a th:href="@{/admin/test/counseling/reply/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/counseling/reply/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title" >고민 상담 게시판 댓글 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="counselingList">
-                            <a th:href="@{/admin/test/store/reply/list}">   <!-- 이동할 페이지 url 작성 -->
+                            <a th:href="@{/admin/store/reply/list}">   <!-- 이동할 페이지 url 작성 -->
                                 <span class="title" >자유 상점 게시판 댓글 관리</span>
                             </a>
                         </li>
@@ -144,12 +144,12 @@
                     </a>
                     <ul>
                         <li class="  " data-title="inquiryList">
-                            <a th:href="@{/admin/test/inquiry/list}">
+                            <a th:href="@{/admin/inquiry/list}">
                                 <span class="title">문의 관리</span>
                             </a>
                         </li>
                         <li class="  " data-title="noticeList">  <!-- a태그의 클래스명이 li 태그 클래스에 출가된다. 활성화 된 클래스 표시(active) -->
-                            <a th:href="@{/admin/test/notice/edit}">  <!-- 해당 페이지의 목록에 active가 들어가야지 음영처리가된다. -->
+                            <a th:href="@{/admin/notice/edit}">  <!-- 해당 페이지의 목록에 active가 들어가야지 음영처리가된다. -->
                                 <span class="title">공지사항 작성</span>
                             </a>
                         </li>
@@ -164,7 +164,7 @@
                     </a>
                     <ul>
                         <li class="  " data-title="configPopup">
-                            <a th:href="@{/admin/test/banner/list}">
+                            <a th:href="@{/admin/banner/list}">
                                 <span class="title">배너 관리</span>
                             </a>
                         </li>
@@ -178,9 +178,5 @@
 </body>
 <script src="https://code.jquery.com/jquery-3.6.1.min.js"></script>
 <script>
-    // $('.title').on("click", function (e) {
-    //     e.preventDefault();
-    //     location.href = "/main/main";
-    // });
 </script>
 </html>
