+++ src/main/resources/templates/app/admin/bannerEdit.html
@@ -3,6 +3,7 @@
 <html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="utf-8">
+    <title>NEOS | 배너 추가</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="">
     <meta name="author" content="">
@@ -118,14 +119,14 @@
                     <div class="col-md-12">
                         <div class="card">
                             <div class="card-body">
-                                <form th:action="@{/admin/test/banner/edit}" th:object="${bannerDTO}" name="bannerEdit" id="dof" method="post" class="form-horizontal form-validate" role="form" target="hidden_frame" enctype="multipart/form-data">
+                                <form th:action="@{/admin/banner/edit}" th:object="${bannerDTO}" name="bannerEdit" id="dof" method="post" class="form-horizontal form-validate" role="form" target="hidden_frame" enctype="multipart/form-data">
                                     <input type="hidden" th:field="*{bannerUrl}" id="bannerUrl">
                                     <input type="hidden" name="popup_idx" id="popup_idx" value="0">
                                     <input type="hidden" name="img_name" id="img_name" value="">
                                     <div class="form-group">
                                         <label class="col-sm-2 control-label">제목</label>
                                         <div class="col-sm-6">
-                                            <input th:field="*{bannerTitle}" type="text" class="form-control" id="title" placeholder="관리용 제목(사용자에겐 표시되지 않습니다)" value="">
+                                            <input th:field="*{bannerTitle}" type="text" class="form-control bannerName" id="title" placeholder="관리용 제목(사용자에겐 표시되지 않습니다)" value="">
                                             <div class="form-control-line"></div>
                                         </div>
                                     </div>
@@ -195,21 +196,6 @@
 </body>
 <script src="https://code.jquery.com/jquery-3.6.1.min.js"></script>
 <script>
-    // base.html
-
-    // 문서로드 완료 후 자바 스크립트 진행
-    // document.addEventListener('DOMContentLoaded', function() {
-    //     const alertBox = document.querySelector('#alertBox');
-    //     const contentArea = document.querySelector('#content');
-    //
-    //     alertBox.classList.remove('alert__wrap-fixed');
-    //     alertBox.classList.add('alert__wrap');
-    //     contentArea.prepend(alertBox);
-    // });
-    //
-    // $(function(){
-    //     SADMIN_MENU_ALERT.openAlert();
-    // });
 
     // 좌측 슬라이드 배너 작동
     $('.gui-folder').click(function(){
@@ -217,49 +203,6 @@
         $(this).parent().children('.expanded').not(this).removeClass('expanded');
     });
 
-    // 값 저장
-    // $(function () {
-    //     CONFIG_POPUP.init({"idx":0,"site_code":null,"unit_code":null,"type":"PL","subject":null,"show_device":[""],"show_page":null,"show_page_list":null,"launch_type":null,"options":{"mobile_position":{"location":{"top":100}}},"content_type":null,"img":null,"content_html":null,"link_url":null,"link_target":null,"start_time":"2022-10-06 00:22","end_time":"2022-10-06 00:22","write_time":null});
-    //
-    //     $('#start_time').datetimepicker({
-    //         dayViewHeaderFormat: 'YYYY MMMM',
-    //         locale: getMomentLangCode(ADMIN_LANG_CODE),
-    //         icons: {
-    //             time: 'zmdi zmdi-time',
-    //             date: 'zmdi zmdi-calendar',
-    //             up: 'zmdi zmdi-chevron-up',
-    //             down: 'zmdi zmdi-chevron-down',
-    //             previous: 'zmdi zmdi-chevron-left',
-    //             next: 'zmdi zmdi-chevron-right',
-    //             today: 'glyphicon glyphicon-screenshot',
-    //             clear: 'glyphicon glyphicon-trash'
-    //         },
-    //         format : 'YYYY-MM-DD HH:mm',
-    //     }).on('blur',function(data){
-    //         if($('#start_time').val() > $('#end_time').val()){
-    //             $('#end_time').val($('#start_time').val());
-    //         }
-    //     });
-    //     $('#end_time').datetimepicker({
-    //         dayViewHeaderFormat: 'YYYY MMMM',
-    //         locale: getMomentLangCode(ADMIN_LANG_CODE),
-    //         icons: {
-    //             time: 'zmdi zmdi-time',
-    //             date: 'zmdi zmdi-calendar',
-    //             up: 'zmdi zmdi-chevron-up',
-    //             down: 'zmdi zmdi-chevron-down',
-    //             previous: 'zmdi zmdi-chevron-left',
-    //             next: 'zmdi zmdi-chevron-right',
-    //             today: 'glyphicon glyphicon-screenshot',
-    //             clear: 'glyphicon glyphicon-trash'
-    //         },
-    //         format : 'YYYY-MM-DD HH:mm',
-    //     }).on('blur',function(data){
-    //         if($('#start_time').val() > $('#end_time').val()){
-    //             $('#end_time').val($('#start_time').val());
-    //         }
-    //     });
-    // });
 
     function removeImg(mode){
         if(mode=='icon'){
@@ -307,32 +250,6 @@
         p.addEventListener("focusout",check_width_validation);
     })
 
-    // 이미지 추가
-    // 업로드한 파일은 input객체의 프로퍼티에 존재한다.
-    // const file = document.querySelector("input[type='file']");
-    // // 사용자가 업로드한 파일에 따라 배경 이미지가 변경되어야 하는 객체
-    // const thumbnail = document.querySelector("label[for='attach'] div");
-    // // change : 해당 객체에 변화가 있을 때(파일이 업로드 되었을 때)
-    // file.addEventListener("change", function(e){
-    //
-    //     // 업로드한 파일을 읽어올 객체
-    //     var reader = new FileReader();
-    //     // 해당 파일의 정보를 readAsDataURL에 전달하면
-    //     // 업로드된 파일의 암호화된 URL정보를 얻을 수 있다.
-    //     reader.readAsDataURL(e.target.files[0]);
-    //     // byte단위로 읽어온 뒤 전부 다 읽으면 load 이벤트가 발생된다.
-    //     reader.onload = function(e){
-    //         // e객체에 읽은 파일의 정보들이 들어있다.
-    //         // console.log(e.target.result);
-    //         let url = e.target.result;
-    //         // 이미지 파일인지 아닌지 검사하여 이미지 파일이 아닐 경우 원래 attach.png 이미지로 변경
-    //         if(url.includes('image')){
-    //             thumbnail.style.backgroundImage = "url('" + url + "')";
-    //         }else{
-    //             thumbnail.style.backgroundImage = "url('img/attach.png')";
-    //         }
-    //     }
-    // });
 
     // 이미지 추가
     // 업로드한 파일은 input객체의 프로퍼티에 존재한다.
@@ -357,6 +274,7 @@
         }
     });
 
+    // 이미지가 바뀔 때 작동되어 DB에 담길 필드에 경로 저장
     $("input[type='file']").on("change", function(){
         let bannerImageFile = $("input[type='file']").get(0).files[0];
         let formData = new FormData();
@@ -377,7 +295,20 @@
         });
     });
 
+
+    // 유효성 검사
     function send() {
+        if($('.bannerName').val() == ""){
+            alert("배너 이름을 추가해주세요!")
+        }else if($("#bannerUrl").val() == "") {
+            alert("배너 이미지를 추가해주세요!");
+        }else {
+            sendData();
+        }
+    }
+
+    // 서브밋
+    function sendData() {
 
         bannerEdit.submit();
     }
