+++ src/main/resources/templates/app/admin/counComment.html
@@ -2,6 +2,7 @@
 <html lang="ko" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="utf-8">
+    <title>NEOS | 고민 게시판 댓글 관리</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta name="description" content="">
     <meta name="author" content="">
@@ -190,6 +191,8 @@
                                                 <tbody>
                                                 <th:block th:each="counselingReply:${counselingReplies}" th:object="${counselingReply}">
                                                 <ul class="content _tbody" id="member_item_m20220929a41c742d48942" data-nick="관리자" data-app="">
+                                                    <input type="hidden" class="counselingId" th:href="*{counselingId}">
+                                                    <input type="hidden" class="userId" th:href="*{userId}">
                                                     <li class="check">
                                                         <div class="checkbox checkbox-styled no-margin">
                                                             <label>
@@ -200,9 +203,9 @@
                                                         </div>
                                                     </li>
 
-                                                    <li class="commentNick"><a th:text="*{userNickName}"></a></li><!-- 작성사 상세 페이지 이동 -->
-                                                    <li class="comment" th:text="*{counselingReplyContent}"></li>
-                                                    <li class="commentPost"><a>[고민상담]&nbsp; <a th:text="*{counselingTitle}"></a></a></li> <!-- 작성한 게시글 페이지 이동 -->
+                                                    <li class="commentNick moveUserDetail"><a href="#" th:text="*{userNickName}"></a></li><!-- 작성사 상세 페이지 이동 -->
+                                                    <li class="comment" style="cursor: pointer" th:text="*{counselingReplyContent}"></li>
+                                                    <li class="commentPost"><a href="#">[고민상담]&nbsp; <a href="#" th:text="*{counselingTitle}"></a></a></li> <!-- 작성한 게시글 페이지 이동 -->
                                                     <li class="commentDate" th:text="*{#temporals.format(createdDate, 'yyyy.MM.dd HH:mm')}"></li>
 
                                                     <li class="action">
@@ -249,22 +252,22 @@
                                last=(${start + 4 < counselingReplies.totalPages ? start + 4 : counselingReplies.totalPages})" th:if="${total != 0}">
                         <th:block th:if="${start!=1}">
                             <a class="btn_paging prev screen_out paging"
-                               th:href="${counselingReplies.first} ? '#' :@{/admin/test/counseling/reply/list(page=${counselingReplies.number-1})}"></a>
+                               th:href="${counselingReplies.first} ? '#' :@{/admin/counseling/reply/list(page=${counselingReplies.number-1})}"></a>
                         </th:block>
                         <th:block th:unless="${start!=1}">
                             <a class="btn_paging prev screen_out paging" style="display: none"
-                               th:href="${counselingReplies.first} ? '#' :@{/admin/test/counseling/reply/list(page=${counselingReplies.number-1})}"></a>
+                               th:href="${counselingReplies.first} ? '#' :@{/admin/counseling/reply/list(page=${counselingReplies.number-1})}"></a>
                         </th:block>
                         <th:block th:each="num : ${#numbers.sequence(start, last)}">
-                            <a class="paging" th:text="${num}" th:href="@{/admin/test/counseling/reply/list(page=${num-1})}"></a>
+                            <a class="paging" th:text="${num}" th:href="@{/admin/counseling/reply/list(page=${num-1})}"></a>
                         </th:block>
                         <th:block th:if="${last != endPage}">
                             <a class="btn_paging next screen_out paging"
-                               th:href="${counselingReplies.last} ? '#' : @{/admin/test/counseling/reply/list(page=${counselingReplies.number+1})}"></a>
+                               th:href="${counselingReplies.last} ? '#' : @{/admin/counseling/reply/list(page=${counselingReplies.number+1})}"></a>
                         </th:block>
                         <th:block th:unless="${last != endPage}">
                             <a class="btn_paging next screen_out paging" style="display: none"
-                               th:href="${counselingReplies.last} ? '#' : @{/admin/test/counseling/reply/list(page=${counselingReplies.number+1})}"></a>
+                               th:href="${counselingReplies.last} ? '#' : @{/admin/counseling/reply/list(page=${counselingReplies.number+1})}"></a>
                         </th:block>
                     </th:block>
                 </span>
@@ -329,6 +332,7 @@
         $deleteModal.show();
     }
 
+    // 현재 페이지 css 주기
     pageActive();
 
     function pageActive() {
@@ -342,6 +346,7 @@
         });
     }
 
+    // 체크된 리스트 삭제
     $('#deleteCheck').click(function() {
         var counselingReplyIds = "";
         $(".terms:checked").each(function() {
@@ -351,15 +356,31 @@
                 counselingReplyIds = counselingReplyIds + "," + $(this).next().attr('href');
             }
         });
-        location.href = "/admin/test/counseling/reply/deleteCheck" + "?counselingReplyIds=" + counselingReplyIds;
+        location.href = "/admin/counseling/reply/deleteCheck" + "?counselingReplyIds=" + counselingReplyIds;
     });
 
+    // 하나만 삭제
     $('.deleteBtn').click(function () {
         let counselingReplyId = $(this).next().attr('href');
         modalUp();
         $('#deleteCheck').click(function () {
-            location.href = "/admin/test/counseling/reply/delete" + "?counselingReplyId=" + counselingReplyId;
+            location.href = "/admin/counseling/reply/delete" + "?counselingReplyId=" + counselingReplyId;
         })
+    })
+
+    // 고민 상담 페이지로 이동
+    $('.comment').click(function () {
+        location.href = "/community/counseling/"
+    })
+
+    // 고민 상담 페이지로 이동
+    $('.commentPost').click(function () {
+        location.href = "/community/counseling/"
+    })
+
+    // 유저 정보로 이동
+    $('.moveUserDetail').click(function () {
+        location.href = "/admin/user/detail?userId=" + $(this).closest('.content').find('.userId').attr('href');
     })
